feat: add pet retrieval and fix potential NullPointerExceptions
- Implement getMyPet in MemberService to retrieve linked Pet data
- Create PetDto record for clean API data transfer
- Fix 'updateMember' always returning true by removing redundant null check
- Prevent NullPointerException in 'getMySaving' by using List.of() for empty transactions
- Update MemberService test suite with refined savings calculation logic

+++ Backend/src/main/java/com/petconomy/controller/MemberController.java
@@ -72,12 +72,12 @@ public class MemberController {
     }
 
     @GetMapping("/{id}")
-    public MemberDto getUser(@PathVariable int id) {
+    public MemberDto getUser(@PathVariable Long id) {
         return memberService.getMember(id);
     }
 
     @DeleteMapping("/{id}")
-    public boolean deleteUser(@PathVariable int id) {
+    public boolean deleteUser(@PathVariable Long id) {
         return memberService.deleteMember(id);
     }
